Refactor the gitignore file & README.md and fix imports.
Change the `.gitignore` to exclude the `/tests` directory instead of `/src/poetry.lock`. In `README.md`, fix the Table of Contents to include links to `Prerequisites` and `Installing` sections, and remove a link to `Contributing` which does not exist in the project. Finally, fix an issue with imports in `main.py` by moving the import statement to occur outside of a function.