Add configuration and build files, refactor cross-validation code
This commit adds several configuration files to the project as well as a build_times.csv file containing the times of different index variations. The cross-validation code was refactored and moved to a new directory, `evaluation/cross_validation`. The new code now uses the `ModelEvaluator` and `ModelTuner` classes, which resulted in significant code reduction. The new code can also read different types of csv files for the queries and relevance judgements.